Prep work for notebook uploading.
