+++ notebook.go
@@ -1,0 +1,7 @@
+package main
+
+func runNotebook(c *cli.Context) {
+  // get images
+  // send multipart form data
+  // open set in Notebooks
+}
